# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for sorting wrong problems.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "1.116.126.208:8080",
    "paths": {
...
        "/app/student/get_student_self": {
            "get": {
                "description": "获取自己的用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "获取自己信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "学生id",
                        "name": "Sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Models.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/student/insert_ask_question": {
            "post": {
                "description": "插入提问问题数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "学生提问",
                "parameters": [
                    {
                        "description": "预约教师数据",
                        "name": "OTData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Models.AskQuestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"Insert() success!\",\"data\":500001}",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Res"
                        }
                    }
                }
            }
        },
        "/app/student/insert_order_teacher": {
            "post": {
                "description": "插入预约教师数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "学生插入预约教师",
                "parameters": [
                    {
                        "description": "预约教师数据",
                        "name": "OTData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Models.OrderTeacher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":200,\"msg\":\"Insert() success!\",\"data\":500001}",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Res"
                        }
                    }
                }
            }
        },
        "/app/student/update_password": {
            "put": {
                "description": "编辑一个学生密码信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "编辑学生密码",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "cid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "cname",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sgrade",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sgradeName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sicon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "snickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "soldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/student/update_student": {
            "post": {
                "description": "编辑一个学生信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "编辑学生",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "cid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "cname",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sgrade",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sgradeName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sicon",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "sid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "snickname",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "soldPassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spassword",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "spasswordAgain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "sphone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Controllers.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Services.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Controllers.Res": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {
                    "type": "object"
                },
                "Msg": {
                    "type": "string"
                }
            }
        },
...
        "Models.AskQuestion": {
            "type": "object",
            "properties": {
                "aqid": {
                    "type": "integer"
                },
                "aqisSolved": {
                    "type": "integer"
                },
                "aqremark": {
                    "type": "string"
                },
                "aqtime": {
                    "type": "string"
                },
                "eid": {
                    "type": "integer"
                },
                "sid": {
                    "type": "integer"
                }
            }
        },
...
        "Models.OrderTeacher": {
            "type": "object",
            "properties": {
                "otaddress": {
                    "type": "string"
                },
                "otcontactName": {
                    "type": "string"
                },
                "otcontactPhone": {
                    "type": "string"
                },
                "otcreateTime": {
                    "type": "string"
                },
                "otid": {
                    "type": "integer"
                },
                "otsubject": {
                    "type": "string"
                },
                "ottimeSlot": {
                    "type": "string"
                },
                "otweek": {
                    "type": "string"
                },
                "sid": {
                    "type": "integer"
                }
            }
        },
        "Models.Student": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "sgrade": {
                    "type": "integer"
                },
                "sicon": {
                    "type": "string"
                },
                "sid": {
                    "type": "integer"
                },
                "sname": {
                    "type": "string"
                },
                "snickname": {
                    "type": "string"
                },
                "spassword": {
                    "type": "string"
                },
                "sphone": {
                    "type": "string"
                }
            }
        },
...
        "Services.Student": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer"
                },
                "cname": {
                    "type": "string"
                },
                "sgrade": {
                    "type": "integer"
                },
                "sgradeName": {
                    "type": "string"
                },
                "sicon": {
                    "type": "string"
                },
                "sid": {
                    "type": "integer"
                },
                "sname": {
                    "type": "string"
                },
                "snickname": {
                    "type": "string"
                },
                "soldPassword": {
                    "type": "string"
                },
                "spassword": {
                    "type": "string"
                },
                "spasswordAgain": {
                    "type": "string"
                },
                "sphone": {
                    "type": "string"
                }
            }
        },
...
    }
}